+++ demos/yjs-react-supabase-text-collab/src/components/providers/SystemProvider.tsx
@@ -1,0 +1,32 @@
+import { AppSchema } from '@/library/powersync/AppSchema';
+import { SupabaseConnector } from '@/library/powersync/SupabaseConnector';
+import { PowerSyncContext } from '@journeyapps/powersync-react';
+import { WASQLitePowerSyncDatabaseOpenFactory } from '@journeyapps/powersync-sdk-web';
+import { CircularProgress } from '@mui/material';
+import Logger from 'js-logger';
+import React, { Suspense } from 'react';
+
+const SupabaseContext = React.createContext<SupabaseConnector | null>(null);
+export const useSupabase = () => React.useContext(SupabaseContext);
+
+export const powerSync = new WASQLitePowerSyncDatabaseOpenFactory({
+  dbFilename: 'powersync2.db',
+  schema: AppSchema
+}).getInstance();
+export const connector = new SupabaseConnector();
+powerSync.connect(connector);
+
+Logger.useDefaults();
+Logger.setLevel(Logger.DEBUG);
+
+export const SystemProvider = ({ children }: { children: React.ReactNode }) => {
+  return (
+    <Suspense fallback={<CircularProgress />}>
+      <PowerSyncContext.Provider value={powerSync}>
+        <SupabaseContext.Provider value={connector}>{children}</SupabaseContext.Provider>
+      </PowerSyncContext.Provider>
+    </Suspense>
+  );
+};
+
+export default SystemProvider;