+++ packages/powersync-sdk-web/src/worker/db/SharedWASQLiteDB.worker.ts
@@ -1,23 +1,58 @@
 import '@journeyapps/wa-sqlite';
 
 import * as Comlink from 'comlink';
+import { v4 as uuid } from 'uuid';
+
 import { DBWorkerInterface, _openDB } from './open-db';
+
+/**
+ * Keeps track of open DB connections and the clients which
+ * are using it.
+ */
+type SharedDBWorkerConnection = {
+  clientIds: Set<string>;
+  db: DBWorkerInterface;
+};
 
 const _self: SharedWorkerGlobalScope = self as any;
 
-const DBMap = new Map<string, Promise<DBWorkerInterface>>();
+const DBMap = new Map<string, SharedDBWorkerConnection>();
+const OPEN_DB_LOCK = 'open-wasqlite-db';
 
 const openDB = async (dbFileName: string): Promise<DBWorkerInterface> => {
-  if (!DBMap.has(dbFileName)) {
-    const openPromise = _openDB(dbFileName);
-    DBMap.set(dbFileName, openPromise);
-    openPromise.catch((error) => {
-      // Allow for retries if an error ocurred
-      console.error(error);
-      DBMap.delete(dbFileName);
-    });
-  }
-  return Comlink.proxy(await DBMap.get(dbFileName)!);
+  // Prevent multiple simultaneous opens from causing race conditions
+  return navigator.locks.request(OPEN_DB_LOCK, async () => {
+    const clientId = uuid();
+
+    if (!DBMap.has(dbFileName)) {
+      const clientIds = new Set<string>();
+      const connection = await _openDB(dbFileName);
+      DBMap.set(dbFileName, {
+        clientIds,
+        db: connection
+      });
+    }
+
+    const dbEntry = DBMap.get(dbFileName)!;
+    dbEntry.clientIds.add(clientId);
+    const { db } = dbEntry;
+
+    const wrappedConnection = {
+      ...db,
+      close: Comlink.proxy(() => {
+        const { clientIds } = dbEntry;
+        clientIds.delete(clientId);
+        if (clientIds.size == 0) {
+          console.debug(`Closing connection to ${dbFileName}.`);
+          DBMap.delete(dbFileName);
+          return db.close?.();
+        }
+        console.debug(`Connection to ${dbFileName} not closed yet due to active clients.`);
+      })
+    };
+
+    return Comlink.proxy(wrappedConnection);
+  });
 };
 
 _self.onconnect = function (event: MessageEvent<string>) {
@@ -26,9 +61,9 @@ _self.onconnect = function (event: MessageEvent<string>) {
   Comlink.expose(openDB, port);
 };
 
-addEventListener('beforeunload', () => {
-  Array.from(DBMap.values()).forEach(async (dbPromise) => {
-    const db = await dbPromise;
+addEventListener('unload', () => {
+  Array.from(DBMap.values()).forEach(async (dbConnection) => {
+    const db = await dbConnection.db;
     db.close?.();
   });
 });