+++ demos/angular-supabase-todolist/CHANGELOG.md
@@ -1,5 +1,13 @@
 # PowerSync Angular Demo
 
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies [1aed928]
+- Updated dependencies [aede9e7]
+  - @journeyapps/powersync-sdk-web@0.3.0
+
 ## 0.0.1
 
 ### Patch Changes